コンポーネントとapiの関数の分離
